+++ website/views/login_user.py
@@ -1,0 +1,40 @@
+# bring in the magic
+from django.contrib.auth import login, authenticate
+from django.http import HttpResponse, HttpResponseRedirect
+from django.shortcuts import render
+from django.template import RequestContext
+
+
+def login_user(request):
+    '''
+    purpose: Handles the creation of a new user for authentication
+
+    author: steve brownlee
+
+    args: request -- The full HTTP request object
+
+    returns: render index view or error if invalid login
+    '''
+
+    # Obtain the context for the user's request.
+    context = RequestContext(request)
+
+    # If the request is a HTTP POST, try to pull out the relevant information.
+    if request.method == 'POST':
+
+        # Use the built-in authenticate method to verify
+        username = request.POST['username']
+        password = request.POST['password']
+        authenticated_user = authenticate(username=username, password=password)
+
+        # If authentication was successful, log the user in
+        if authenticated_user is not None:
+            login(request=request, user=authenticated_user)
+            return HttpResponseRedirect('/')
+
+        else:
+            # Bad login details were provided. So we can't log the user in.
+            print("Invalid login details: {}, {}".format(username, password))
+            return HttpResponse("Invalid login details supplied.")
+
+    return render(request, 'login.html', {}, context)
