+++ website/views/view_account.py
@@ -1,0 +1,6 @@
+from django.shortcuts import render
+
+def view_account(request):
+  if request.method == 'GET':
+    template_name = 'account/view_account.html'
+    return render(request, template_name)
